+++ R/ip_report.R
@@ -5,7 +5,7 @@
 #' @param ip a valid IPv4 address in dotted quad notation; String; Required 
 #' @param \dots Additional arguments passed to \code{\link{virustotal_GET}}.
 #' 
-#' @return named list with the following items: 
+#' @return named list with the following potential items: 
 #' \code{undetected_referrer_samples, detected_downloaded_samples, detected_referrer_samples, 
 #' undetected_downloaded_samples, detected_urls, undetected_downloaded_samples, response_code, as_owner, verbose_msg, country, 
 #' undetected_referrer_samples, detected_communicating_samples, resolutions, undetected_communicating_samples, asn}
@@ -26,7 +26,7 @@
 ip_report <- function(ip = NULL, ...) {
 
     if (!is.character(ip)) {
-        stop("Must specify IP")
+        stop("Must specify a valid IP.\n")
     }
 
     params <- list(ip = ip)