+++ R/scan_url.R
@@ -1,5 +1,8 @@
 #' Submit URL for scanning
 #' 
+#' Submit a URL for scanning. Returns a data.frame with \code{scan_id} which can be used to 
+#' fetch the report using \code{\link{url_report}}
+#' 
 #' @param url url; string; required
 #' @param \dots Additional arguments passed to \code{\link{virustotal_POST}}.
 #' 
@@ -22,11 +25,11 @@
 scan_url <- function(url = NULL, ...) {
 
     if (!is.character(url)) {
-        stop("Must specify url")
+        stop("Must specify a valid url.\n")
     }
 
     res    <- virustotal_POST(path="url/scan", query = list(url = url), ...)
     
-    as.data.frame(do.call(cbind, res))
+    as.data.frame(res)
 }